refactor(frontend): concurrent rendering in React 18
- use the new `createRoot` API in React 18.
- refactor the map layers to avoid an error in React Map GL when the map first mounts.
