+++ frontend/src/map/MapView.tsx
@@ -1,18 +1,10 @@
-import { Suspense, useCallback, useEffect } from 'react';
-import {
-  atom,
-  useRecoilState,
-  useRecoilValue,
-  useResetRecoilState,
-  useSetRecoilState,
-} from 'recoil';
+import { Suspense, useCallback } from 'react';
+import { useSetRecoilState } from 'recoil';
 
-import { mapViewStateState } from '../state/map-view/map-view-state';
-import { BoundingBox } from 'lib/bounding-box';
 import { BaseMap } from 'lib/data-map/BaseMap';
 import { DataMap } from 'lib/data-map/DataMap';
 import { DataMapTooltip } from 'lib/data-map/DataMapTooltip';
-import { MapBoundsFitter } from 'lib/map/MapBoundsFitter';
+import { MapBoundsFitter, mapFitBoundsState } from 'lib/map/MapBoundsFitter';
 import { MapHud } from 'lib/map/hud/MapHud';
 import { MapHudRegion } from 'lib/map/hud/MapHudRegion';
 import {
@@ -25,22 +17,12 @@ import { PlaceSearchResult } from 'lib/map/place-search/use-place-search';
 import { ErrorBoundary } from 'lib/react/ErrorBoundary';
 import { withProps } from 'lib/react/with-props';
 
-import { interactionGroupsState } from 'state/layers/interaction-groups';
-import { viewLayersFlatState } from 'state/layers/view-layers-flat';
-import { useSaveViewLayers, viewLayersParamsState } from 'state/layers/view-layers-params';
 import { globalStyleVariables } from '../theme';
 import { useIsMobile } from '../use-is-mobile';
 
 import { MapLayerSelection } from './layers/MapLayerSelection';
-import { backgroundState, showLabelsState } from './layers/layers-state';
 import { MapLegend } from './legend/MapLegend';
 import { TooltipContent } from './tooltip/TooltipContent';
-import { useBasemapStyle } from './use-basemap-style';
-
-export const mapFitBoundsState = atom<BoundingBox>({
-  key: 'mapFitBoundsState',
-  default: null,
-});
 
 const AppPlaceSearch = () => {
   const setFitBounds = useSetRecoilState(mapFitBoundsState);
@@ -114,40 +96,12 @@ const MapHudMobileLayout = () => {
 };
 
 const MapViewContent = () => {
-  const [viewState, setViewState] = useRecoilState(mapViewStateState);
-  const background = useRecoilValue(backgroundState);
-  const showLabels = useRecoilValue(showLabelsState);
-  const viewLayers = useRecoilValue(viewLayersFlatState);
-  const saveViewLayers = useSaveViewLayers();
-  const { mapStyle, firstLabelId } = useBasemapStyle(background, showLabels);
-
-  useEffect(() => {
-    saveViewLayers(viewLayers);
-  }, [saveViewLayers, viewLayers]);
-
-  const viewLayersParams = useRecoilValue(viewLayersParamsState);
-
-  const interactionGroups = useRecoilValue(interactionGroupsState);
-
-  const fitBounds = useRecoilValue(mapFitBoundsState);
-
-  const resetFitBounds = useResetRecoilState(mapFitBoundsState);
-  useEffect(() => {
-    // reset map fit bounds whenever MapView is mounted
-    resetFitBounds();
-  }, [resetFitBounds]);
-
   const isMobile = useIsMobile();
 
   return (
-    <BaseMap mapStyle={mapStyle} viewState={viewState} onViewState={setViewState}>
-      <DataMap
-        beforeId={firstLabelId}
-        viewLayers={viewLayers}
-        viewLayersParams={viewLayersParams}
-        interactionGroups={interactionGroups}
-      />
-      <MapBoundsFitter boundingBox={fitBounds} />
+    <BaseMap>
+      <DataMap />
+      <MapBoundsFitter />
       <DataMapTooltip>
         <TooltipContent />
       </DataMapTooltip>
